Handle literals; generalise parsing
Pass literals as pairs of markers that delimit the content in the
process buffer.  At a later point we'll have to flush old content from
the process buffer.

I believe that the IMAP protocol is general enough that I can treat most
of it as a sequence of strings, atoms and lists, the exception being the
constructs that are followed by resp-text.  Thus I don't need a special
case for every possible response in bic--parse; on the other hand, the
client gets slightly more raw data.  Let's see how this works out.