Disable undo in connection buffers
Now that we delete data from connection buffers, the undo list grows.
We don't need undo anyway, so we should just disable it from the start.

Eventually, the name of the connection buffer should be changed to start
with a space, so that undo is disabled automatically, but I'm not ready
to do that in this stage of development.