Accept :deactivate in state :existing
This occurs when we cannot connect to a server that we know about
because the connection times out.  This means relatively long periods of
time where we would previously not be able to just deactivate the
connection.  Now we can stop its reconnection cycle until we know it
makes sense to retry.