Try to detect connections silently lost
If we have sent a command to the IMAP server, and haven't received a
single byte during the last 30 seconds, consider the connection to be
dead.

This seems to happen when I close the laptop lid: the connection is in
fact unusable when I open it again, but the OS doesn't tell us.  Let's
try to detect this anyway.